+++ src/pymbus/exceptions.py
@@ -1,6 +1,13 @@
+"""Meter-Bus (M-Bus) exception classes."""
+
+
 class MBusError(Exception):
     """Meter-Bus Base Error."""
 
 
 class MBusLengthError(MBusError):
-    """Invalid length/size error."""
+    """Invalid Data Length/Size Error."""
+
+
+class MBusValidationError(MBusError):
+    """Data Validation Error."""
