+++ src/pymbus/telegrams/base.py
@@ -1,10 +1,18 @@
-from collections.abc import Generator, Iterable
+"""Basic Telegram objects module.
 
-# from typing import Self
-from pymbus.exceptions import MBusError
+Glossary:
+
+- a byte is an integer witing the range(0, 256)
+- an instance of the TelegramField class is a wrapper over a byte
+- a telegram byte refers to the "int | TelegramField"
+"""
+
+from collections.abc import Iterable, Iterator
+
+from pymbus.exceptions import MBusValidationError
 
 
-def validate_byte(nbr: int) -> int:
+def _validate_byte(nbr: int) -> int:
     """Validates an integer number to be a byte.
 
     In Python, a byte must be in range(0, 256).
@@ -22,7 +30,8 @@ def validate_byte(nbr: int) -> int:
     try:
         bytes([nbr])
     except ValueError as e:
-        raise MBusError from e
+        msg = f"{nbr} is not a valid byte"
+        raise MBusValidationError(msg) from e
 
     return nbr
 
@@ -30,17 +39,17 @@ def validate_byte(nbr: int) -> int:
 class TelegramField:
     """The base "Field" class.
 
-    It is a base wrapper for a byte value.
+    It is a base wrapper over a byte value (validation is ensured).
     A Field is a part of blocks, frames and other Telegram containers.
     """
 
-    def __init__(self, byte: int):
-        self._byte = validate_byte(byte)
+    def __init__(self, byte: int) -> None:
+        self._byte = _validate_byte(byte)
 
-    def __eq__(self, other) -> bool:
+    def __eq__(self, other: object) -> bool:
         sbyte = self.byte
         if isinstance(other, TelegramField):
-            return sbyte == other.byte
+            other = other.byte
         return sbyte == other
 
     def __repr__(self) -> str:
@@ -54,28 +63,16 @@ class TelegramField:
         return self._byte
 
 
-TelegramBytesType = bytes | bytearray | Iterable[int | TelegramField]
+TelegramByteType = int | TelegramField
 
 
-def parse_byte(byte_like: int | TelegramField) -> int:
-    """Return the byte value from the `byte_like` argument.
-
-    If byte_like is a TelegramField (TF),
-    then its byte property get called.
-
-    Parameters
-    ----------
-    byte_like: int | TelegramField (TF)
-        either a byte-like integer or a TF as a byte wrapper
-
-    Returns
-    -------
-    int
-    """
-
-    if isinstance(byte_like, TelegramField):
-        return byte_like.byte
-    return byte_like
+def _convert_to_telegram_fields(
+    ibytes: Iterable[TelegramByteType],
+) -> list[TelegramField]:
+    return [
+        ibyte if isinstance(ibyte, TelegramField) else TelegramField(ibyte)
+        for ibyte in ibytes
+    ]
 
 
 class TelegramContainer:
@@ -86,50 +83,53 @@ class TelegramContainer:
     """
 
     @classmethod
-    def from_hexstring(cls, hexstr: str):
+    def from_hexstring(cls, hexstr: str) -> "TelegramContainer":
         """Return a class instance from a hexadecimal string."""
 
-        barr = bytearray.fromhex(hexstr)
-        return cls(barr)
+        try:
+            return cls(bytearray.fromhex(hexstr))
+        except ValueError as e:
+            raise MBusValidationError(str(e)) from e
 
     @classmethod
-    def from_integers(cls, ints: Iterable[int]):
+    def from_integers(cls, ints: Iterable[int]) -> "TelegramContainer":
         """Return a class instance from a sequence of integers."""
 
-        barr = bytearray(iter(ints))
-        return cls(barr)
+        try:
+            return cls(bytearray(iter(ints)))
+        except ValueError as e:
+            raise MBusValidationError(str(e)) from e
 
-    def __init__(self, ibytes: TelegramBytesType) -> None:
-        self._fields = []
-        for ib in ibytes:
-            field = ib if isinstance(ib, TelegramField) else TelegramField(ib)
-            self._fields.append(field)
+    def __init__(
+        self, ibytes: None | Iterable[TelegramByteType] = None
+    ) -> None:
+        self._fields = _convert_to_telegram_fields(ibytes or [])
 
-    def __eq__(self, other) -> bool:
-        sfields = self.fields
+    def __eq__(self, other: object) -> bool:
+        sfields = self._fields
         if isinstance(other, TelegramContainer):
-            return sfields == other.fields
+            other = other._fields
         return sfields == other
 
-    def __getitem__(self, idx: int) -> TelegramField:
-        return self._fields[idx]
+    def __getitem__(
+        self, key: int | slice
+    ) -> "TelegramField | TelegramContainer":
+        if isinstance(key, int):
+            return self._fields[key]
+        return type(self)(self._fields[key])
 
-    def __iter__(self) -> Generator[None, None, TelegramField]:
-        yield from self.fields
+    def __iter__(self) -> Iterator[TelegramField]:
+        yield from self._fields
 
     def __len__(self) -> int:
-        return len(self.fields)
+        return len(self._fields)
 
     def __repr__(self) -> str:
         cls_name = type(self).__name__
-        return f"{cls_name}(fields={self.fields})"
+        return f"{cls_name}(ibytes={self._fields})"
 
     def __str__(self) -> str:
         return str(list(self))
 
-    @property
-    def fields(self) -> list[TelegramField]:
-        return self._fields
-
     def as_bytes(self) -> bytes:
-        return bytes(field.byte for field in self.fields)
+        return bytes(field.byte for field in self._fields)
